requirements.txt
PyYAML ~= 6.0

# Install custom-conf in editable/development mode
-e .
